+++ front-end/tests/tests.rs
@@ -4,6 +4,7 @@ mod config;
 mod encoding;
 mod file_discovery;
 mod help;
+#[cfg(windows)]
 mod idempotence;
 mod io_error;
 mod modes;
